Adding build-replica.sh
I've set up the script for building a replica outside of docker-entrypoint.sh - it needs to be the entrypoint instead of docker-entrypoint.sh for the time being but other than that it's able to initiate a base backup. Removing the custom entrypoint for db1 means we will need to modify pg_hba.conf and postgresql.conf some other way.